+++ src/__generated__/FairDetailShowsQuery.graphql.ts
@@ -144,6 +144,7 @@ fragment LocationMap_location on Location {
   address
   address_2
   postal_code
+  summary
   coordinates {
     lat
     lng
@@ -363,7 +364,7 @@ return {
   "kind": "Request",
   "operationKind": "query",
   "name": "FairDetailShowsQuery",
-  "id": "46b2db28051cad863890f63b5af30bd9",
+  "id": "4ce30624864ce54f3d7cb43936580874",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -673,6 +674,13 @@ return {
                 "kind": "ScalarField",
                 "alias": null,
                 "name": "postal_code",
+                "args": null,
+                "storageKey": null
+              },
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "summary",
                 "args": null,
                 "storageKey": null
               },