+++ src/lib/Scenes/Fair/Components/FairHeader/index.tsx
@@ -68,7 +68,7 @@ export class FairHeader extends React.Component<Props, State> {
       .map(i => {
         if (i.node.partner && i.node.partner.profile && i.node.partner.profile.name) {
           return {
-            href: "show/" + i.node.id,
+            href: "booth/" + i.node.id,
             name: i.node.partner.profile.name,
           }
         }